ASU-314: Separate stickyNavigation to its own minified JS-file.
