+++ public/themes/custom/asuntotuotanto/asuntotuotanto.libraries.yml
@@ -1,3 +1,7 @@
+sticky-navigation:
+  js:
+    dist/js/stickyNavigation.min.js: {}
+
 global-styling:
   version: 1.x
   css:
