+++ public/themes/custom/asuntotuotanto/templates/content/node--apartment--full.html.twig
@@ -71,6 +71,8 @@
  */
 #}
 
+{{ attach_library('asuntotuotanto/sticky-navigation') }}
+
 {%
   set classes = [
   'node--type-' ~ node.bundle|clean_class,
