+++ public/themes/custom/asuntotuotanto/webpack.config.build.js
@@ -1,14 +1,17 @@
-const merge = require('webpack-merge');
-const webpackConfig = require('./webpack.config');
-const MinifyPlugin = require('babel-minify-webpack-plugin');
+const merge = require("webpack-merge");
+const MinifyPlugin = require("babel-minify-webpack-plugin");
+const webpackConfig = require("./webpack.config");
 
 module.exports = merge(webpackConfig, {
-  mode: 'production',
-  devtool: '',
+  mode: "production",
+  devtool: "",
   plugins: [
-    new MinifyPlugin({}, {
-      comments: false,
-      sourceMap: '',
-    })
-  ]
+    new MinifyPlugin(
+      {},
+      {
+        comments: false,
+        sourceMap: "",
+      }
+    ),
+  ],
 });
